# Proposal: Optimize `func goodName(name string) bool` in `text/template`