## Description

I have been benchmarking an alternative implementations of the `goodName` function, currently used in the `text/template` package. The function validates function identifiers in templates but may benefit from optimization for performance. My results suggest that an alternative implementation performs consistently better in various scenarios.

## Background

The `goodName` function is used to validate function identifiers in templates. The function is defined as follows in [Go 1.23.4](https://cs.opensource.google/go/go/+/refs/tags/go1.23.4:src/text/template/funcs.go;l=127-141):

```go
func goodName(name string) bool {
    if name == "" {
        return false
    }
    for i, r := range name {
        switch {
        case r == '_':
        case i == 0 && !unicode.IsLetter(r):
            return false
        case !unicode.IsLetter(r) && !unicode.IsDigit(r):
            return false
        }
    }
    return true
}
```

The function returns true if the input string is a valid identifier and false otherwise. The function is called for each function identifier in the template.

## Proposal

I propose to optimize the `goodName` function by reordering and simplifying the checks. The proposed implementation is as follows:

```go
func goodName(name string) bool {
    if name == "" {
        return false
    }
    for i, r := range name {
        switch {
        case unicode.IsLetter(r):
        case r == '_':
        case i > 0 && unicode.IsDigit(r):
        default:
            return false
        }
    }
    return true
}
```

- For the most common case, a letter, the function do only one check.
- The next most common case is an underscore, which is checked next.
- Finally, the function checks for a digit.

The proposed implementation is simpler and more efficient than the current implementation. The proposed implementation has been benchmarked and shown to be faster than the current implementation except in some "strange" cases, where a valid identifier do not sontain any letter.

## Benchmark

Here is the benchmark code used to compare the two implementations:

```go
package goodname

import (
    "testing"
    "unicode"
)

func goodNameOld(name string) bool {
    if name == "" {
        return false
    }
    for i, r := range name {
        switch {
        case r == '_':
        case i == 0 && !unicode.IsLetter(r):
            return false
        case !unicode.IsLetter(r) && !unicode.IsDigit(r):
            return false
        }
    }
    return true
}

func goodNameNew(name string) bool {
    if name == "" {
        return false
    }
    for i, r := range name {
        switch {
        case unicode.IsLetter(r):
        case r == '_':
        case i > 0 && unicode.IsDigit(r):
        default:
            return false
        }
    }
    return true
}

var goodNames = []string{
    "validname",
    "valid_name",
    "_valid_name",
    "validname1",
    "valid_name1",
    "_valid_name1",
    "validname_2025",
}

func BenchmarkGoodNameOld_valid(b *testing.B) {
    for i := 0; i < b.N; i++ {
        for _, name := range goodNames {
            _ = goodNameOld(name)
        }
    }
}

func BenchmarkGoodNameNew_valid(b *testing.B) {
    for i := 0; i < b.N; i++ {
        for _, name := range goodNames {
            _ = goodNameNew(name)
        }
    }
}

var badNames = []string{
    "1nvalid",
    "i-valid",
    "in-valid",
    "invali-d",
}

func BenchmarkGoodNameOld_invalid(b *testing.B) {
    for i := 0; i < b.N; i++ {
        for _, name := range badNames {
            _ = goodNameOld(name)
        }
    }
}

func BenchmarkGoodNameNew_invalid(b *testing.B) {
    for i := 0; i < b.N; i++ {
        for _, name := range badNames {
            _ = goodNameNew(name)
        }
    }
}

var strangeNames = []string{
    "________",
    "_1234567",
}

func BenchmarkGoodNameOld_strange(b *testing.B) {
    for i := 0; i < b.N; i++ {
        for _, name := range strangeNames {
            _ = goodNameOld(name)
        }
    }
}

func BenchmarkGoodNameNew_strange(b *testing.B) {
    for i := 0; i < b.N; i++ {
        for _, name := range strangeNames {
            _ = goodNameNew(name)
        }
    }
}
```

And here are the results of the benchmark:

```shell
go test -bench=.
goos: windows
goarch: amd64
pkg: kpym.github.com/gobench/goodname
cpu: Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz
BenchmarkGoodNameOld_valid-8             7354932               148.9 ns/op
BenchmarkGoodNameNew_valid-8             9289075               118.4 ns/op
BenchmarkGoodNameOld_invalid-8          27421300                37.41 ns/op
BenchmarkGoodNameNew_invalid-8          35162773                28.62 ns/op
BenchmarkGoodNameOld_strange-8          34371643                29.46 ns/op
BenchmarkGoodNameNew_strange-8          29400090                34.70 ns/op
PASS
ok      kpym.github.com/gobench/goodname        6.893s
```

We can see that the new implementation is around 20% faster for valid identifiers and around 30% faster for invalid identifiers (which is not important because wa need a fast implementation for valid identifiers). 

The new implementation is slower only for the odd cases where the identifier is valid but does not contain any letter.

## Compatibility

The proposed implementation is compatible with the current implementation. The function signature remains the same, and the behavior is unchanged. The proposed implementation is more efficient and should be a drop-in replacement for the current implementation.

## Documentation

By the way, I was not able to find any documentation about what is a valid function identifier in the text/template package. It would be nice to have a documentation about this.

_A valid function name is a sequence of letters, digits, and underscores, but must start with a letter or an underscore._

## Conclusion

I propose to optimize the `goodName` function in the text/template package. The proposed implementation is simpler and more efficient than the current implementation. The proposed implementation has been benchmarked and shown to be faster than the current implementation except in some "strange" cases, where a valid identifier do not sontain any letter.